Moved axios interceptor to separate utils folder to remove redundancy
